requirements.txt
fastapi
uvicorn
snowflake-connector-python
pandas
scikit-learn
streamlit
google-genai
python-dotenv